moved Formatter and Key to utils
