+++ leptos_i18n_macro/src/utils/mod.rs
@@ -1,6 +1,8 @@
 use quote::{quote, TokenStreamExt};
 use syn::Token;
 
+pub mod formatter;
+pub mod key;
 pub mod scoped;
 
 pub enum Keys {
